Canbios para manejar mas aplicaciones y preparar el proyecto para nuevos proyectos
