+++ TecunRecolecta/Recolecta/settings.py
@@ -31,6 +31,7 @@ ALLOWED_HOSTS = ['procitcluster.com', '3.101.143.28', 'localhost', '127.0.0.1']
 
 INSTALLED_APPS = [
     'clientes',
+    'energiaplantas',
     'django.contrib.admin',
     'django.contrib.auth',
     'django.contrib.contenttypes',
@@ -76,7 +77,7 @@ WSGI_APPLICATION = 'Recolecta.wsgi.application'
 DATABASES = {
     'default': {
         'ENGINE': 'django.db.backends.postgresql',
-        'NAME': 'clientes',
+        'NAME': 'recolecta',
         'USER': 'postgres',
         'PASSWORD': 'postgres',
         'HOST': 'localhost',
@@ -133,3 +134,8 @@ DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'
 
 LOGIN_REDIRECT_URL = 'index'
 LOGOUT_REDIRECT_URL = 'index'
+
+from django.contrib.messages import constants as message_constants
+MESSAGE_TAGS = {
+    message_constants.ERROR: 'danger',  # Para usar con Bootstrap
+}
