+++ TecunRecolecta/clientes/templates/clientes.html
@@ -10,6 +10,7 @@
     <title>Tecun Recolecta</title>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet"
           integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/js/bootstrap.bundle.min.js"></script>
     <link rel="stylesheet" href="{% static 'main.css' %}">
     <link rel="stylesheet" href="https://unpkg.com/leaflet@1.7.1/dist/leaflet.css" />
     <style>
@@ -18,7 +19,6 @@
             width: 100%;   /* Ancho completo */
             margin-top: 20px; /* Espaciado entre los elementos */        }
     </style>
-
 </head>
 <body>
 <div class="container">
@@ -29,39 +29,26 @@
                 <div class="d-flex align-items-center">
                     <img src="{% static 'images/tecun.jpg' %}" alt="Tecun" class="img-fluid"
                          style="max-width: 50px; margin-right: 10px;">
-                    <h4>Equipos de {{ cliente.nombre_completo }}</h4>
+                    <h4>Clientes</h4>
                 </div>
-                <a href="{% url 'opciones' %}">Regresar</a>
+                <a href="{% url 'index' %}">Regresar</a>
             </div>
 
-            <form action="{% url 'newe' %}" method="POST" class="card card-body">
+            <form action="{% url 'newcli' %}" method="POST" class="card card-body">
                 {% csrf_token %}
-                <input type="hidden" name="cliente_id" value="{{ cliente.id }}">
-                <label for="descripcion">Descripcion:</label>
-                <input type="text" name="descripcion" placeholder="Descripcion de Equipo" class="form-control mb-2">
-                <label for="Horometro">Horometro:</label>
-                <input type="text" name="Horometro" placeholder="Horometro" class="form-control mb-2">
-                <label for="potencia">Potencia:</label>
-                <input type="text" name="potencia" placeholder="Potencia" class="form-control mb-2">
 
-
-                {{ mmotors.as_p }}
-                {{ mgeneradors.as_p }}
-                {{ mcontroladors.as_p }}
-                {{ mmodulos.as_p }}
-                {{ usos.as_p }}
-
-                <label for="observaciones">Observaciones:</label>
-                <textarea name="observaciones" rows="3" placeholder="Observaciones"
-                          class="form-control mb-2"></textarea>
-                <div>
-                <input type="checkbox" id="grabarubi" name="grabarubi">
-                </div>
-                <label for="grabaubi">grabar ubicacion?</label>
+                <label for="nombre_completo">Nombre Completo:</label>
+                <input type="text" name="nombre_completo" placeholder="Nombre Completo" class="form-control mb-2" required=True>
+                <h4>Ubicacion Geografica</h4>
+                <p></p>
                 <label for="lat">Latitud:</label>
                 <input type="text" id="lat" name="lat">
                 <label for="lon">Longitud:</label>
                 <input type="text" id="lon" name="lon">
+                <div class="form-check">
+                    <input type="checkbox" class="form-check-input" id="grabarubi" name="grabarubi">
+                     <label class="form-check-label" for="grabarubi">Grabar ubicación?</label>
+                </div>
                 <div class="col-md-4" id="map"></div>
                 <script src="https://unpkg.com/leaflet@1.7.1/dist/leaflet.js"></script>
                 <script>
@@ -72,8 +59,7 @@
                              document.getElementById("lat").value = latitude
                              const longitude = position.coords.longitude;
                              document.getElementById("lon").value = longitude
-
-                    // Muestra la ubicación en el mapa
+                             // Muestra la ubicación en el mapa
                     var map = L.map('map').setView([latitude, longitude], 13);
                     L.tileLayer('https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png', {
                         attribution: '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors'
@@ -89,28 +75,21 @@
 
 
                 </script>
-
-
                 <button class="btn btn-success">Guardar</button>
             </form>
-
-            {% for equipo in equipos %}
+            {% for cliente in clientes %}
             <div class="card card-body text-black mt-2 d-flex flex-row justify-content-between">
                 <div>
-                    <h4>{{equipo.descripcion}}</h4>
-                    {{equipo.marcamotor}} | {{equipo.marcacontrolador}} | {{equipo.marcagenerador}} |
-                    {{equipo.marcamodulo}} | {{equipo.Horometro}} | {{equipo.Potencia}} |
-                    {{equipo.latitude}} | {{equipo.longitude}} | {{equipo.uso}} |
-                    {{equipo.observaciones}}
+                    <h4>{{cliente.nombre_completo}}</h4>
+                    <h6>{{cliente.cli_id}} | {{cliente.latitude}} | {{cliente.longitude}} </h6>
                 </div>
-                <form action="{% url 'dele' equipo.id %}" method="POST">
+                <form action="{% url 'delcli' cliente.id %}" method="POST">
                     {% csrf_token %}
-                    <input type="hidden" name="cliente_id" value="{{ cliente.id }}">
                     <button class="btn btn-danger btn-sm">Borrar</button>
                 </form>
             </div>
             {% empty %}
-            <h1 class="display-4 text-center">No ha registrado Equipos</h1>
+            <h1 class="display-4 text-center">No ha registrado Clientes</h1>
             {% endfor %}
 
             {% else %}
