+++ TecunRecolecta/energiaplantas/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class EnergiaplantasConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'energiaplantas'
