+++ TecunRecolecta/energiaplantas/templates/cliugep.html
@@ -31,7 +31,7 @@
                          style="max-width: 50px; margin-right: 10px;">
                     <h4>Selecciona Cliente</h4>
                 </div>
-                <a href="{% url 'index' %}">Regresar</a>
+                <a href="{% url 'indexep' %}">Regresar</a>
             </div>
             <form action="{% url 'uptuc' %}" method="POST" class="card card-body">
                 {% csrf_token %}
@@ -77,7 +77,7 @@
                     Registra Ubicacion e Ir a Equipos
                 </button>
                 <p></p>
-                <button class="btn btn-secondary" onclick="document.getElementById('action_type').value='no_registra_ubicacion';">
+                <button class="btn btn-info" onclick="document.getElementById('action_type').value='no_registra_ubicacion';">
                   Ir a Equipos
                 </button>
             </form>
