+++ TecunRecolecta/energiaplantas/templates/equiposep.html
@@ -1,0 +1,125 @@
+{% load static %}
+
+<!doctype html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport"
+          content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <title>Tecun Recolecta</title>
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet"
+          integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
+    <link rel="stylesheet" href="{% static 'main.css' %}">
+    <link rel="stylesheet" href="https://unpkg.com/leaflet@1.7.1/dist/leaflet.css" />
+    <style>
+        #map {
+            height: 400px;
+            width: 100%;   /* Ancho completo */
+            margin-top: 20px; /* Espaciado entre los elementos */        }
+    </style>
+
+</head>
+<body>
+<div class="container">
+    <div class="row">
+        <div class="col-md-4 offset-md-4">
+            {% if user.is_authenticated %}
+            <div class="card card-body text-black mt-2 d-flex flex-row justify-content-between">
+                <div class="d-flex align-items-center">
+                    <img src="{% static 'images/tecun.jpg' %}" alt="Tecun" class="img-fluid"
+                         style="max-width: 50px; margin-right: 10px;">
+                    <h4>Equipos de {{ cliente.nombre_completo }}</h4>
+                </div>
+                <a href="{% url 'cliugep' %}">Regresar</a>
+            </div>
+
+            <form action="{% url 'newep' %}" method="POST" class="card card-body">
+                {% csrf_token %}
+                <input type="hidden" name="cliente_id" value="{{ cliente.id }}">
+                <label for="descripcion">Descripcion:</label>
+                <input type="text" name="descripcion" placeholder="Descripcion de Equipo" class="form-control mb-2" required=True>
+                <label for="Horometro">Horometro:</label>
+                <input type="number" name="Horometro" placeholder="Horometro" class="form-control mb-2">
+                <label for="potencia">Potencia:</label>
+                <input type="number" name="potencia" placeholder="Potencia" class="form-control mb-2">
+
+
+                {{ mmotors.as_p }}
+                {{ mgeneradors.as_p }}
+                {{ mcontroladors.as_p }}
+                {{ mmodulos.as_p }}
+                {{ usos.as_p }}
+
+                <label for="observaciones">Observaciones:</label>
+                <textarea name="observaciones" rows="3" placeholder="Observaciones"
+                          class="form-control mb-2"></textarea>
+                <div class="form-check">
+                    <input type="checkbox" class="form-check-input" id="grabarubi" name="grabarubi">
+                     <label class="form-check-label" for="grabarubi">Grabar ubicación?</label>
+                </div>
+                <label for="lat">Latitud:</label>
+                <input type="text" id="lat" name="lat">
+                <label for="lon">Longitud:</label>
+                <input type="text" id="lon" name="lon">
+                <div class="col-md-4" id="map"></div>
+                <script src="https://unpkg.com/leaflet@1.7.1/dist/leaflet.js"></script>
+                <script>
+
+                    if (navigator.geolocation) {
+                         navigator.geolocation.getCurrentPosition(function(position) {
+                             const latitude = position.coords.latitude;
+                             document.getElementById("lat").value = latitude
+                             const longitude = position.coords.longitude;
+                             document.getElementById("lon").value = longitude
+
+                    // Muestra la ubicación en el mapa
+                    var map = L.map('map').setView([latitude, longitude], 13);
+                    L.tileLayer('https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png', {
+                        attribution: '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors'
+                    }).addTo(map);
+                    L.marker([latitude, longitude]).addTo(map)
+                        .bindPopup('Tu ubicación actual').openPopup();
+
+
+                         });
+                     } else {
+                         alert('Geolocalización no es soportada por este navegador.');
+                     }
+
+
+                </script>
+
+
+                <button class="btn btn-success">Guardar</button>
+            </form>
+
+            {% for equipo in equipos %}
+            <div class="card card-body text-black mt-2 d-flex flex-row justify-content-between">
+                <div>
+                    <h4>{{equipo.descripcion}}</h4>
+                    {{equipo.marcamotor}} | {{equipo.marcacontrolador}} | {{equipo.marcagenerador}} |
+                    {{equipo.marcamodulo}} | {{equipo.Horometro}} | {{equipo.Potencia}} |
+                    {{equipo.latitude}} | {{equipo.longitude}} | {{equipo.uso}} |
+                    {{equipo.observaciones}}
+                </div>
+                <form action="{% url 'delep' equipo.id %}" method="POST">
+                    {% csrf_token %}
+                    <input type="hidden" name="cliente_id" value="{{ cliente.id }}">
+                    <button class="btn btn-danger btn-sm">Borrar</button>
+                </form>
+            </div>
+            {% empty %}
+            <h1 class="display-4 text-center">No ha registrado Equipos</h1>
+            {% endfor %}
+
+            {% else %}
+            <h2> No esta autenticado </h2>
+            <a href="{% url 'login' %}">Ingresar</a>
+            {% endif %}
+
+        </div>
+    </div>
+</div>
+</body>
+</html>
