+++ TecunRecolecta/energiaplantas/templates/marcasgenerador.html
@@ -23,7 +23,7 @@
                          style="max-width: 50px; margin-right: 10px;">
                     <h4>Marcas Generador</h4>
                 </div>
-                <a href="{% url 'index' %}">Regresar</a>
+                <a href="{% url 'indexep' %}">Regresar</a>
             </div>
 
             <form action="{% url 'newmg' %}" method="POST" class="card card-body">
