+++ TecunRecolecta/energiaplantas/templates/marcasmotor.html
@@ -10,6 +10,7 @@
     <title>Tecun Recolecta</title>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet"
           integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/js/bootstrap.bundle.min.js"></script>
     <link rel="stylesheet" href="{% static 'main.css' %}">
 </head>
 <body>
@@ -21,29 +22,30 @@
                 <div class="d-flex align-items-center">
                     <img src="{% static 'images/tecun.jpg' %}" alt="Tecun" class="img-fluid"
                          style="max-width: 50px; margin-right: 10px;">
-                    <h4>Usos de Equipos</h4>
+                    <h4>Marcas Motor</h4>
                 </div>
-                <a href="{% url 'index' %}">Regresar</a>
+                <a href="{% url 'indexep' %}">Regresar</a>
             </div>
 
-            <form action="{% url 'newu' %}" method="POST" class="card card-body">
+            <form action="{% url 'newmm' %}" method="POST" class="card card-body">
                 {% csrf_token %}
 
-                <label for="descripcion">Uso Equipo:</label>
-                <input type="text" name="descripcion" placeholder="Descripcion Uso Equipo" class="form-control mb-2">
-                <button class="btn btn-success">Guardar</button></form>
-            {% for uso in usos %}
+                <label for="descripcion">Marca Motor:</label>
+                <input type="text" name="descripcion" placeholder="Descripcion Marca Motor" class="form-control mb-2">
+                <button class="btn btn-success">Guardar</button>
+            </form>
+            {% for mmotor in mmotors %}
             <div class="card card-body text-black mt-2 d-flex flex-row justify-content-between">
                 <div>
-                    <h4>{{uso.descripcion}}</h4>
+                    <h4>{{mmotor.descripcion}}</h4>
                 </div>
-                <form action="{% url 'delu' uso.id %}" method="POST">
+                <form action="{% url 'delmm' mmotor.id %}" method="POST">
                     {% csrf_token %}
                     <button class="btn btn-danger btn-sm">Borrar</button>
                 </form>
             </div>
             {% empty %}
-            <h1 class="display-4 text-center">No ha registrado Usos de Equipos</h1>
+            <h1 class="display-4 text-center">No ha registrado Marcas Motor</h1>
             {% endfor %}
 
             {% else %}
